Store original image shape when using --nodata
For large images such as satellite images it is often good to only store
the polygons and labels without replicating the actual image data. This
is already solved by the --nodata option. However, in order to create
rasterized masks with the same dimension as the image data it is useful
to know the size of the original image. This could of course be solved
by adding logic involing the original image file, but it's a little more
practical to just store it along with the json file.